refactor api for modular routes
